mysite/requirements.txt
asgiref==3.7.2
autopep8==2.3.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
Django==5.0.9
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
Markdown==3.7
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
requests==2.32.3
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.3.0
whitenoise==6.9.0
